==> code/4. NN_Likelihood_encoding.ipynb <==
#Import necessary libraries
# To write a Python 2/3 compatible codebase, the first step is to add this line to the top of each module
from __future__ import division, print_function, unicode_literals

from keras.models import Sequential
from keras.layers import Dense, Activation
from keras import optimizers


import numpy as np
import pandas as pd
import random as rn
import tensorflow as tf
np.random.seed(3456)
rn.seed(3456)
session_conf =tf.compat.v1.ConfigProto(intra_op_parallelism_threads=1,inter_op_parallelism_threads=1)

from keras import backend as K
tf.random.set_seed(3456)
sess = tf.compat.v1.Session(graph=tf.compat.v1.get_default_graph(),config=session_conf)
#K.set_session(sess)
tf.compat.v1.keras.backend.set_session(sess)
import warnings
warnings.filterwarnings("ignore")

#settings for binary classifier NN
NEpochs = 1000
BatchSize=250
Optimizer=optimizers.RMSprop(lr=0.01)

#set the directory
import os
os.chdir(r"C:\Users\banva\Desktop\ML Project\Mayur Kumar")
os.getcwd()

# Read in the data
calibrationdata=  pd.read_csv('calibration_erik.csv',sep=',',header=0,quotechar='"')
validationdata=  pd.read_csv('validation_erik.csv',sep=',',header=0,quotechar='"')

cal_Target = calibrationdata.loc[:,["click"]]
val_Target = validationdata.loc[:,["click"]]

cal_Target.click=cal_Target.click.astype('category',copy=False)
val_Target.click=val_Target.click.astype('category',copy=False)

cal_Target.dtypes
val_Target.dtypes

vars=list(calibrationdata)[2:30]
vars

calibrationdata_x =calibrationdata.loc[:,vars]

validationdata_x =validationdata.loc[:,vars]

# loop to change each column to category type
for col in vars:
    validationdata_x[col] = validationdata_x[col].astype('float64',copy=False)

calibrationdata_x.dtypes

validationdata_x.dtypes

XTr= np.array(calibrationdata_x)
XVal=np.array(validationdata_x)
BCNN.add(Dense(units=4,activation="relu",use_bias=True))

YTr= np.array(cal_Target) 
YVal=np.array(val_Target)

BCNN = Sequential()

BCNN.add(Dense(units=4,input_shape=(XTr.shape[1],),activation="relu",use_bias=True))
BCNN.add(Dense(units=4,activation="relu",use_bias=True))
BCNN.add(Dense(units=4,activation="relu",use_bias=True))
BCNN.add(Dense(units=4,activation="relu",use_bias=True))
BCNN.add(Dense(units=1,activation="sigmoid",use_bias=True))

BCNN.compile(loss='binary_crossentropy', optimizer=Optimizer,metrics=['binary_crossentropy'])


#%% Fit NN Model,p
FitHist = BCNN.fit(XTr,YTreoschs=NEpochs,batch_size=BatchSize,verbose=0'])
print("Number of Epochs = "+tr(len(FitHist.history['binary_crossentropy)))
FitHist.history['binary_crossentropy'][-1]
FitHist.history['binay_crossentropy'][-10:-1]

#%% Make Predictionsoo
YHatTr = BCNN.predict(XTr,batch_size=XTr.shape[0]) # NteNot scaled, s ont nceessary to undo.
YHatTr = YHatTr.reshape((YHatTr.shape[0]),)

YHatVal = BCNN.predict(XVal,batch_size=XVal.shape[0])
YHatVal = YHatVal.reshape((YHatVal.shape[0]),) 

TrOutDF = pd.DataFrame(data={ 'YHatTr': YHatTr, 'YTr':YTr}})

ValOutDF = pd.DataFrame(data={ 'YHatVal': YHatVal , 'YVal':YVal })

TrOutDF.to_csv('TrYHatFromBCNN.csv',sep=',',na_rep="NA",header=True,index=False)
ValOutDF.to_csv('ValYHatFromBCNN.csv',sep=',',na_rep="NA",header=True,index=False)
